UI working properly

+++ src/main/java/FairGrounds/Domain/ExpertiseProfile.java
@@ -29,7 +29,7 @@ public class ExpertiseProfile implements ExpertiseProfileDTO{
     private Application application;
 
     @Column (name = "years")
-    private long years;
+    private int years;
 
     public Expertise getExpertise() {
         return expertise;
@@ -55,11 +55,11 @@ public class ExpertiseProfile implements ExpertiseProfileDTO{
         this.application = application;
     }
 
-    public long getYears() {
+    public int getYears() {
         return years;
     }
 
-    public void setYears(long years) {
+    public void setYears(int years) {
         this.years = years;
     }
 }
